Added more views and second window for camera.  Cassettes and Pick in main window.  Next, view for camera controls, a board view?
