+++ MainWindow.xaml
@@ -6,21 +6,15 @@
         xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
         xmlns:local="clr-namespace:Picky" 
         mc:Ignorable="d" FontSize="12" 
-        Height="1040" Width="1700" ResizeMode="CanResize"
+        Height="800" Width="800" ResizeMode="CanResize"
         WindowStyle="SingleBorderWindow" WindowStartupLocation="CenterScreen" Background="Transparent">
 
     <Window.Resources>
-        <DataTemplate DataType="{x:Type local:MachineViewModel}">
-            <local:MachineView/>
-        </DataTemplate>
         <DataTemplate DataType="{x:Type local:CassetteViewModel}">
             <local:CassetteView/>
         </DataTemplate>
-        <DataTemplate DataType="{x:Type local:ControlViewModel}">
-            <local:ControlView/>
-        </DataTemplate>
-        <DataTemplate DataType="{x:Type local:CameraViewModel}">
-            <local:CameraView/>
+        <DataTemplate DataType="{x:Type local:PickListViewModel}">
+            <local:PickListView/>
         </DataTemplate>
     </Window.Resources>
 
@@ -38,14 +32,9 @@
                     <ColumnDefinition Width="auto"/>
                     <ColumnDefinition Width="auto"/>
                 </Grid.ColumnDefinitions>
-                <!--Commands/Settings-->
-                <Grid Grid.Column="0" Name="ctrlView" Margin="10" VerticalAlignment="Top"/>
-                <!-- Adding MachineView -->
-                <Grid Grid.Column="1" Name="mView"/>
-                
             </Grid>
-            <!-- Adding CassetteView -->
-            <Grid Name="cView" Grid.Row="1"  Grid.Column="1" Margin="10 10 10 10" />
+            <Grid Name="lView" Grid.Row="0"  Grid.Column="0" Margin="10 10 10 10" />
+            <Grid Name="cView" Grid.Row="1"  Grid.Column="0" Margin="10 10 10 10" />
         </Grid>
     </Border>
 </Window>
