+++ PickListView.xaml
@@ -1,0 +1,80 @@
+﻿<UserControl x:Class="Picky.PickListView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Picky"
+             d:DataContext="{d:DesignInstance Type=local:PickListViewModel}"
+             xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="600">
+
+    
+
+    <Expander IsExpanded="True" Padding="10" >
+        <Expander.Header>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="auto"/>
+                    <ColumnDefinition Width="*"/>
+                </Grid.ColumnDefinitions>
+                <TextBlock Grid.Column="0" Text="Project Pick List" Margin="0,0,0,0" HorizontalAlignment="Left" TextAlignment="Left" />
+            </Grid>
+        </Expander.Header>
+        <Expander.Content>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"/>
+                </Grid.ColumnDefinitions>
+                <Grid Grid.Column="0" Margin="5 -10 5 5" >
+                    <!-- Placement 2 Rows -->
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="45" />
+                        <RowDefinition Height="300" />
+                    </Grid.RowDefinitions>
+                    <Grid Grid.Row="0" Margin="10 0 10 0" >
+                        <!-- Placement 1 Row - Title and Basic Controls -->
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="auto"/>
+                            <ColumnDefinition Width="*"/>
+                        </Grid.ColumnDefinitions>
+                        <TextBlock Grid.Column="0" Text="" Margin="10,0,0,0" HorizontalAlignment="Left" VerticalAlignment="Center"  TextAlignment="Left" />
+                        <ui:Button Grid.Column="1" Content="Open..." Command="{Binding OpenPickListCommand}" Icon="Folder32"  Height="30" Margin="0,0,0,0" HorizontalAlignment="Right"/>
+                    </Grid>
+                    <ui:DataGrid  x:Name="PickListGrid" Grid.Row="1" HorizontalAlignment="Stretch" SelectedItem="{Binding selectedPickListPart}" 
+                          ItemsSource="{Binding PickList}" AutoGenerateColumns="False" CanUserSortColumns="False"
+                          HorizontalContentAlignment="Stretch" SelectionChanged="PickListSelectionChanged"
+                          VerticalScrollBarVisibility="Auto" SelectionMode="Single" SelectionUnit="FullRow" IsReadOnly="True" >
+                        <!-- Placement 2nd Row - Data Grid -->
+                        <DataGrid.Columns>
+                            <DataGridTextColumn Binding="{Binding Designator}" Header="Designator" Width="80"/>
+                            <DataGridTextColumn Binding="{Binding Description}" Header="Description" Width="auto"/>
+                            <DataGridTextColumn Binding="{Binding CenterX}" Header="Center-X" Width="70"/>
+                            <DataGridTextColumn Binding="{Binding CenterY}" Header="Center-Y" Width="70"/>
+                            <DataGridTextColumn Binding="{Binding Rotation}" Header="Rotation" Width="70"/>
+                            <DataGridTextColumn Binding="{Binding cassette.name}" Header="Cassette" Width="70"/>
+                        </DataGrid.Columns>
+                        <ui:DataGrid.RowStyle>
+                            <Style TargetType="DataGridRow">
+                                <Style.Triggers>
+                                    <Trigger Property="IsSelected" Value="True">
+                                        <Setter Property="Background" Value="#80B9EE"/>
+                                        <Setter Property="Foreground" Value="Black"/>
+                                    </Trigger>
+                                </Style.Triggers>
+                            </Style>
+                        </ui:DataGrid.RowStyle>
+                        <DataGrid.ContextMenu>
+                            <ContextMenu>
+                                <MenuItem Header="Go to Location"  Command="{Binding GoToPartLocationCommand}"/>
+                                <MenuItem Header="Place at Location"  Command="{Binding feeder.PlacePartAtLocationCommand}"/>
+                                <Separator Margin="2"/>
+                                <MenuItem Header="Add to Cassette" Command="{Binding AddPartToCassetteCommand}" />
+                            </ContextMenu>
+                        </DataGrid.ContextMenu>
+                    </ui:DataGrid>
+                </Grid>
+            </Grid>
+        </Expander.Content>
+    </Expander>
+</UserControl>
