+++ Picky.csproj
@@ -141,6 +141,9 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="BoardView.xaml.cs">
+      <DependentUpon>BoardView.xaml</DependentUpon>
+    </Compile>
     <Compile Include="CalibrationModel.cs" />
     <Compile Include="CameraView.xaml.cs">
       <DependentUpon>CameraView.xaml</DependentUpon>
@@ -168,11 +171,22 @@
       <DependentUpon>MachineView.xaml</DependentUpon>
     </Compile>
     <Compile Include="MachineViewModel.cs" />
+    <Compile Include="PickListView.xaml.cs">
+      <DependentUpon>PickListView.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="PickListViewModel.cs" />
     <Compile Include="PickModel.cs" />
+    <Compile Include="PickView.xaml.cs">
+      <DependentUpon>PickView.xaml</DependentUpon>
+    </Compile>
     <Compile Include="RelayInterface.cs" />
     <Compile Include="Part.cs" />
     <Compile Include="RelayCommand.cs" />
     <Compile Include="SerialInterface.cs" />
+    <Page Include="BoardView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="CameraView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -205,6 +219,14 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Page Include="PickListView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="PickView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs">
